requirements.txt
aiohttp >=3.7
fsspec >=0.8.5
h5netcdf >=0.8.1
intake-xarray
intake>2
pydap
siphon
tqdm
